[autocommit] Commit and push all org-mode files automatically
All org-mode files in the ~/org directory are automatically committed to
git every 5 minutes. If the network is reachable, changes are also
pulled and pushed from/to Gitlab.

The implementation as an infinite loop managed by runit instead of as a
cron job is ugly, but I didn't see a good way to manage the crontab with
stow and it has the benefit that all output is written to a (rotated) logfile.